soy idiota equisdé
Fixed SignalFragment.HasValuesAt && temporary offset calculator

Implemented SignalFragment.GetChildrenTouching ISignalContainer Overload

+++ Signal Breaker Unity Project/Assets/Scripts/SignalHierarchy/Interfaces/MathShit.cs
@@ -13,7 +13,7 @@
 		//if a loopLength is provided calculate overflow
 		int loopedPosition = position % (int)loopLength;
 		//ensure loop-relative position is in the positive range
-		if (loopedPosition < 0) { loopedPosition += loopLength; }
+		if (loopedPosition < 0) { loopedPosition += (int)loopLength; }
 
 		return loopedPosition;
